+++ src/main/java/core/basesyntax/service/impl/StorageService.java
@@ -1,5 +1,5 @@
 package core.basesyntax.service.impl;
 
 public interface StorageService {
-    void operateSupply(String forFruit, Double quantity);
+    void operateSupply(String forFruit, Integer quantity);
 }
